requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-crispy-forms==2.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
pip==23.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
PyPDF2==3.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==67.6.1
sqlparse==0.4.4
urllib3==2.0.4